Split resource loader from mod initialiazer.
