+++ src/main/java/fr/estecka/variantscit/VariantsCitMod.java
@@ -3,31 +3,22 @@ package fr.estecka.variantscit;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.model.loading.v1.ModelLoadingPlugin;
 import net.fabricmc.fabric.api.client.model.loading.v1.PreparableModelLoadingPlugin;
-import net.fabricmc.fabric.api.client.model.loading.v1.PreparableModelLoadingPlugin.DataLoader;
 import net.minecraft.client.item.ModelPredicateProviderRegistry;
 import net.minecraft.item.Item;
 import net.minecraft.item.Items;
 import net.minecraft.registry.Registries;
-import net.minecraft.resource.Resource;
-import net.minecraft.resource.ResourceManager;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.JsonHelper;
-import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.Executor;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import com.google.gson.JsonElement;
-import com.mojang.serialization.JsonOps;
 
 import fr.estecka.variantscit.modules.*;
 
 
 public class VariantsCitMod
-implements ClientModInitializer, PreparableModelLoadingPlugin<Map<Item,VariantManager>>, DataLoader<Map<Item,VariantManager>>
+implements ClientModInitializer, PreparableModelLoadingPlugin<Map<Item,VariantManager>>
 {
 	static public final String MODID = "variants-cit";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MODID);
@@ -40,7 +31,7 @@ implements ClientModInitializer, PreparableModelLoadingPlugin<Map<Item,VariantMa
 
 	@Override
 	public void onInitializeClient(){
-		PreparableModelLoadingPlugin.register(this, this);
+		PreparableModelLoadingPlugin.register(new ModuleLoader(), this);
 		ModuleRegistry.Register(Identifier.ofVanilla("stored_enchantments"), new EnchantedBookModule());
 		ModuleRegistry.Register(Identifier.ofVanilla("axolotl_variant"), new AxolotlBucketModule());
 		ModuleRegistry.Register(Identifier.ofVanilla("potion_effect"), new PotionEffectModule());
@@ -65,60 +56,4 @@ implements ClientModInitializer, PreparableModelLoadingPlugin<Map<Item,VariantMa
 		}
 	}
 
-	@Override
-	public CompletableFuture<Map<Item,VariantManager>> load(ResourceManager resourceManager, Executor executor){
-		return CompletableFuture.supplyAsync(()->Reload(resourceManager), executor);
-	}
-
-	private Map<Item,VariantManager> Reload(ResourceManager manager){
-		Map<Item,VariantManager> result = new HashMap<>();
-
-		for (Map.Entry<Identifier, Resource> entry : manager.findResources("variant-cits/item", id->id.getPath().endsWith(".json")).entrySet()){
-			Identifier resourceId = entry.getKey();
-			Item item = ItemFromModule(resourceId);
-			if (item == null)
-				continue;
-
-			JsonElement json;
-			VariantManager module;
-			try {
-				json = JsonHelper.deserialize(entry.getValue().getReader());
-			}
-			catch (IOException e){
-				VariantsCitMod.LOGGER.error("{}", e);
-				continue;
-			}
-
-			var dataResult = ModuleDefinition.CODEC.decoder().decode(JsonOps.INSTANCE, json);
-			if (dataResult.isError()){
-				VariantsCitMod.LOGGER.error("Error in cit module {}: {}", resourceId, dataResult.error().get().message());
-				continue;
-			}
-
-			try {
-				module = ModuleRegistry.CreateManager(dataResult.getOrThrow().getFirst());
-			}
-			catch (IllegalArgumentException e){
-				VariantsCitMod.LOGGER.error("Error building cit module of type {}: {}", dataResult.getPartialOrThrow().getFirst().type(), e);
-				continue;
-			}
-
-			module.ReloadVariants(manager);
-			result.put(item, module);
-		}
-
-		return result;
-	}
-
-	static private @Nullable Item ItemFromModule(Identifier resourceId){
-		String path = resourceId.getPath();
-		path = path.substring("variant-cits/item".length() + 1, path.length() - ".json".length());
-
-		Identifier itemId = Identifier.of(resourceId.getNamespace(), path);
-
-		if (Registries.ITEM.containsId(itemId))
-			return Registries.ITEM.get(itemId);
-		else
-			return null;
-	}
 }
